+++ docs/changes.md
@@ -1,4 +1,10 @@
 # 最新动态
+* 2018/11/13
+  * 修改agg/agge为nanovg后端时linecap的BUG。
+
+* 2018/11/12
+  * 增加js demos。
+
 * 2018/11/11
   * 完善ProgressCircle控件。
   * base/time.h改名为base/time\_now.h，避免和系统time.h冲突。